pyproject.toml
[tool.poetry]
name = "2022-python-lab"
version = "0.1.0"
description = "Experiments in modern Python"
authors = ["RonquilloAeon"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
xpresso = "^0.33.0"
ulid-py = "^1.1.0"
uvicorn = "^0.17.6"
pydantic = "^1.9.0"
aioredis = "^2.0.1"
msgpack = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
requests = "^2.27.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
